Standardized figure output in `LongShortTrackComparingDisplayFunctions` for all functions except the interactive one (the first one). Added standardized defer_render and save_figure arguments to the functions.
Also changed `_make_pho_jonathan_batch_plots` to return a `MatplotlibRenderPlots` object instead of a dict. Note that the function that uses it `_display_batch_pho_jonathan_replay_firing_rate_comparison` previously had all of the saving code in a different file that saved after calling the function, but now it saves internally (and only saves .png where it previously saved .pdf and .png as options).
I think .pdf could potentially be added if `perform_write_to_file` were updated, but there's that weird thing where the plot call for .pdf saving has to be wrapped in that matplotlib pdf context manager, so that doesn't fit the current workflow so well.